Refactor key binding logic
